Pull out common reducer.js logic to media/common
